refactor endpoint validation network rules
